tastyai/pyproject.toml
[project]
name = "tastyai"
version = "0.1.0"
description = ""
authors = [
    {name = "Matheus Eidt",email = "matheus@mx2.unisc.br"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "transformers (>=4.49.0,<5.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "openai (>=1.64.0,<2.0.0)",
    "streamlit (>=1.42.2,<2.0.0)",
    "googletrans (>=4.0.2,<5.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "pip (>=25.0.1,<26.0.0)",
    "setuptools (>=75.8.1,<76.0.0)",
    "wheel (>=0.45.1,<0.46.0)",
    "spacy (>=3.8.4,<4.0.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "langchain (>=0.3.19,<0.4.0)",
    "langchain-openai (>=0.3.7,<0.4.0)",
    "pillow (>=11.1.0,<12.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
